ctx/requirements.txt
asn1crypto==1.4.0
asttokens==2.0.5
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coincurve==13.0.0
colorama==0.4.4
cytoolz==0.11.0
devtools==0.8.0
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-typing==2.2.2
eth-utils==1.10.0
executing==0.8.2
halo==0.0.31
iconsdk==1.3.6
idna==3.2
log-symbols==0.0.14
multipledispatch==0.6.0
ping3==3.0.1
pycparser==2.20
pycryptodome==3.10.1
PyYAML==5.4.1
requests==2.26.0
requests-mock==1.8.0
simple-term-menu==1.3.0
six==1.16.0
socket-request==0.1.32
spinners==0.0.24
termcolor==1.1.0
toolz==0.11.1
urllib3==1.26.6
aiofiles==0.7.0
xxhash==2.0.2
argparse
boto3==1.18.36
ffcount==0.1.9
cryptography==35.0.0
